+++ jest.config.json
@@ -13,7 +13,8 @@
     "**/*.(t|j)s",
     "!./main.ts",
     "!./app.module.ts",
-    "!./config/logger.config.ts"
+    "!./config/logger.config.ts",
+    "!./auth/oauth.controller.ts"
   ],
   "coverageDirectory": "../coverage",
   "testEnvironment": "node",
