+++ src/auth/auth.controller.spec.ts
@@ -139,78 +139,6 @@ describe('AuthController', () => {
     });
   });
 
-  describe('googleLoginCallback', () => {
-    let authController: AuthController;
-    const user = createOauthUser({ providerType: OauthProvider.GOOGLE });
-    const jwtToken = createJwtToken({});
-    beforeEach(async () => {
-      const module = await getTestingModule(
-        new AuthServiceStub(user, jwtToken),
-      );
-      authController = module.get<AuthController>(AuthController);
-    });
-
-    it('기존에 가입한 유저인 경우 해당 유저 객체를 반환', async () => {
-      // given
-      const oauthPayload = OauthPayloadDto.create({
-        email: EMAIL,
-        nickname: NICKNAME,
-        providerType: OauthProvider.GOOGLE,
-        providerId: PROVIDER_ID,
-      });
-
-      // when
-      const jwtSignResultDto = await authController.googleLoginCallback(
-        oauthPayload,
-      );
-
-      // then
-      expect(jwtSignResultDto.userId).toBe(USER_ID);
-      expect(jwtSignResultDto.accessToken).toBeTruthy();
-      expect(jwtSignResultDto.refreshToken).toBeTruthy();
-    });
-
-    it('신규 사용자인 경우 새로운 유저를 생성하여 일부 정보를 반환', async () => {
-      // given
-      const oauthPayload = OauthPayloadDto.create({
-        email: 'new@email.com',
-        nickname: 'new_nickname',
-        providerType: OauthProvider.GOOGLE,
-        providerId: 'new_oauth123',
-      });
-
-      // when
-      const jwtSignResultDto = await authController.googleLoginCallback(
-        oauthPayload,
-      );
-
-      // then
-      expect(jwtSignResultDto.userId).toBeTruthy();
-      expect(jwtSignResultDto.accessToken).toBeTruthy();
-      expect(jwtSignResultDto.refreshToken).toBeTruthy();
-    });
-
-    it('다른 로그인 방식으로 가입한 유저이면 UnauthorizedException 발생', async () => {
-      // given
-      const oauthPayload = OauthPayloadDto.create({
-        email: EMAIL,
-        nickname: NICKNAME,
-        providerType: OauthProvider.LOCAL,
-        providerId: PROVIDER_ID,
-      });
-
-      // when
-      // then
-      await expect(
-        authController.googleLoginCallback(oauthPayload),
-      ).rejects.toThrow(
-        new UnauthorizedException(
-          AuthErrorMessage.OAUTH_PROVIDER_UNMATCHED[oauthPayload.providerType],
-        ),
-      );
-    });
-  });
-
   describe('rotateToken', () => {
     let authController: AuthController;
     const user = createUser({});
